requirements.txt
requests
unittest2
websocket-client
